chore: tidy registry tests in to separate __tests__ dir
