+++ .gitignore
@@ -17,6 +17,7 @@ __pycache__
 coverage/
 
 # IDE files
+.kiro/
 .vscode/
 .idea/
 *.swp
