+++ src/registry/registry.ts
@@ -194,166 +194,3 @@ export abstract class Registry<T, I> {
     return undefined
   }
 }
-
-// Unit tests
-if (import.meta.vitest) {
-  const { describe, it, expect, beforeEach, vi } = import.meta.vitest
-
-  // A concrete implementation of the abstract Registry for testing purposes
-  class TestRegistry extends Registry<string, number> {
-    private nextId = 1
-
-    protected generateId(): number {
-      return this.nextId++
-    }
-
-    protected validate(item: string): void {
-      if (item.length === 0) {
-        throw new ValidationError('Item cannot be an empty string.')
-      }
-    }
-  }
-
-  describe('Error Classes', () => {
-    it('ItemNotFoundError should have the correct name and message', () => {
-      const error = new ItemNotFoundError(123)
-      expect(error.name).toBe('ItemNotFoundError')
-      expect(error.message).toBe("Item with id '123' not found")
-    })
-
-    it('DuplicateItemError should have the correct name and message', () => {
-      const error = new DuplicateItemError('abc')
-      expect(error.name).toBe('DuplicateItemError')
-      expect(error.message).toBe("An item with the ID 'abc' already exists.")
-    })
-
-    it('ValidationError should have the correct name and message', () => {
-      const error = new ValidationError('Invalid item')
-      expect(error.name).toBe('ValidationError')
-      expect(error.message).toBe('Invalid item')
-    })
-  })
-
-  describe('Registry', () => {
-    let registry: TestRegistry
-
-    beforeEach(() => {
-      registry = new TestRegistry()
-    })
-
-    it('should register an item and return a new ID', () => {
-      const id = registry.add('test-item')
-      expect(id).toBe(1)
-      expect(registry.get(1)).toBe('test-item')
-    })
-
-    it('should throw DuplicateItemError when registering with an existing ID', () => {
-      // @ts-expect-error - Spying on protected 'generateId' to test duplicate handling.
-      const generateIdSpy = vi.spyOn(registry, 'generateId').mockReturnValue(1)
-      registry.add('test-item') // This will register with ID 1.
-      expect(() => registry.add('another-item')).toThrow(DuplicateItemError)
-      generateIdSpy.mockRestore()
-    })
-
-    it('should deregister an item and return it', () => {
-      const id = registry.add('test-item')
-      const deregisteredItem = registry.remove(id)
-      expect(deregisteredItem).toBe('test-item')
-      expect(registry.get(id)).toBeUndefined()
-    })
-
-    it('should throw ItemNotFoundError when deregistering a non-existent item', () => {
-      expect(() => registry.remove(999)).toThrow(ItemNotFoundError)
-    })
-
-    it('should get an item by its ID', () => {
-      const id = registry.add('test-item')
-      const foundItem = registry.get(id)
-      expect(foundItem).toBe('test-item')
-    })
-
-    it('should return undefined when getting a non-existent item', () => {
-      const foundItem = registry.get(999)
-      expect(foundItem).toBeUndefined()
-    })
-
-    it('should find an item using a predicate', () => {
-      registry.add('item-a')
-      registry.add('item-b')
-      const foundItem = registry.find((item) => item.includes('b'))
-      expect(foundItem).toBe('item-b')
-    })
-
-    it('should return undefined when no item matches the predicate', () => {
-      registry.add('item-a')
-      const foundItem = registry.find((item) => item.includes('c'))
-      expect(foundItem).toBeUndefined()
-    })
-
-    it('should return all keys', () => {
-      registry.add('item-1')
-      registry.add('item-2')
-      expect(registry.keys()).toEqual([1, 2])
-    })
-
-    it('should return all values', () => {
-      registry.add('item-1')
-      registry.add('item-2')
-      expect(registry.values()).toEqual(['item-1', 'item-2'])
-    })
-
-    it('should return all key-value pairs', () => {
-      registry.add('item-1')
-      registry.add('item-2')
-      expect(registry.pairs()).toEqual([
-        [1, 'item-1'],
-        [2, 'item-2'],
-      ])
-    })
-
-    it('should clear all items from the registry', () => {
-      registry.add('item-1')
-      registry.clear()
-      expect(registry.keys()).toEqual([])
-      expect(registry.values()).toEqual([])
-    })
-
-    it('should register multiple items', () => {
-      const ids = registry.addAll(['item-a', 'item-b'])
-      expect(ids).toEqual([1, 2])
-      expect(registry.values()).toEqual(['item-a', 'item-b'])
-    })
-
-    it('should deregister multiple items', () => {
-      const ids = registry.addAll(['item-a', 'item-b', 'item-c'])
-      const deregisteredItems = registry.removeAll([ids[0]!, ids[2]!])
-      expect(deregisteredItems).toEqual(['item-a', 'item-c'])
-      expect(registry.values()).toEqual(['item-b'])
-    })
-
-    it('should find and deregister an item', () => {
-      registry.add('item-a')
-      registry.add('item-b')
-      const deregisteredItem = registry.findRemove((item) => item.includes('a'))
-      expect(deregisteredItem).toBe('item-a')
-      expect(registry.values()).toEqual(['item-b'])
-    })
-
-    it('should return undefined from findRemove if no item matches', () => {
-      const removedItem = registry.findRemove((item) => item.includes('c'))
-      expect(removedItem).toBeUndefined()
-    })
-
-    it('should call the validate method on register', () => {
-      // @ts-expect-error - Spying on protected 'validate' to confirm it is called.
-      const validateSpy = vi.spyOn(registry, 'validate')
-      registry.add('a-valid-item')
-      expect(validateSpy).toHaveBeenCalledWith('a-valid-item')
-      validateSpy.mockRestore()
-    })
-
-    it('should throw a validation error for an invalid item', () => {
-      expect(() => registry.add('')).toThrow(ValidationError)
-    })
-  })
-}
